+++ lib/generators/double_entry/reporting/install/templates/migration.rb
@@ -2,15 +2,15 @@ class CreateDoubleEntryReportingTables < ActiveRecord::Migration<%= migration_ve
   def self.up
     create_table "double_entry_line_aggregates", :force => true do |t|
       t.string     "function",   :limit => 15, :null => false
-      t.string     "account",    :limit => 31, :null => false
-      t.string     "code",       :limit => 47
-      t.string     "scope",      :limit => 23
+      t.string     "account",                  :null => false
+      t.string     "code"
+      t.string     "scope"
       t.integer    "year"
       t.integer    "month"
       t.integer    "week"
       t.integer    "day"
       t.integer    "hour"
-      t.integer    "amount",                   :null => false
+      t.bigint     "amount",                   :null => false
       t.string     "filter"
       t.string     "range_type", :limit => 15, :null => false
       t.timestamps                             :null => false